refactor(types): Move PhotonUpgrades into types.ts.
